Level 3 & 4
Blocked out and added assets to levels 3 and 4
i did not touch the destroyable objects because of the way the script was set up it doesnt handle doing multiple objects under one parent. I woulda needed to adjust Jame's code but as it I won't touch it as the code is not mine.